cgroup: Do not count cpu stats with disabled cpuacct controller
Currently, we do double accounting both for the v1 cpuacct controller
and for the cgroup core in v2. The former is pointless work for users of
unified hierarchy only. We cannot simply give up on accounting when
cpuacct is not mounted (we would miss all charges since boot), so at
least allow users to disabled cpuacct controller explicitly.

TODO measure