psi/cgroup: Deduplicate static_branch PSI check
The check in psi_trigger_create() is redundant (unreachable) to the
check in psi.c:psi_write(). Remove it.
The PSI cgroup files aren't created when the branch is disabled
(css_populate_dir()'s handling of cgroup_psi_files).
To remain on the safe side add a cgroup-specific check in the cgroup
write handlers and fail early (the new check should also be unreachable
today).

No functional change intended.

Signed-off-by: Michal Koutný <mkoutny@suse.com>