cgroup: Disallow mounting v1 hierarchies without controller implementation
The configs that disable some v1 controllers would still allow mounting
them but with no controller-specific files. (Equvalent to named v1
hierarchies.) To achieve behavior consistent with actual out-compilation
of a whole controller, the mounts should treat respective controllers as
non-existent.

Exploit the cgroup_no_v1 mask for the implementation, the effect is that
mounts would fail and produce a message like:
  [ 1450.390260] cgroup: Disabled controller 'memory'

Signed-off-by: Michal Koutný <mkoutny@suse.com>