kernfs: Drop break/unbreak active protection mechanism
The mechanism was needed to allow self-removal of kernfs files.
Since the active_ref is converted to RCU with asynchronous calls, we can
now afford to get rid of the break/unbreak mechanism. IOW, it was needed
to prevent the following deadlock:
	rcu_read_lock()
		break_protection // rcu_read_unlock
		synchronize_rcu()
		unbreak_protection // rcu_read_lock
		// gray zone
	rcu_read_unlock()
But since kernfs file removal/draining uses the asynchronous call_rcu()
that breaks the loop implicitly, we can drop the exemption parentheses.

Signed-off-by: Michal Koutný <mkoutny@suse.com>